requirements.txt
youtube_transcript_api
streamlit
google-generativeai